New feature automatic stream pause.
If you leave the buffer void, when the inputs are collected by the function `collect_inputs`, because the Actor is idle:
the NetworkedController doesn't send to the Server any additional data.

The Server, is now able to figure it out and automatically generates a void buffer. So the simulation
can keep going without any input being sent through internet.